authoptions filename and imports renamed
