+++ app/user/[id]/edit/page.tsx
@@ -2,7 +2,7 @@ import prisma from '@/db';
 import EditUserForm from '../../conponents/edit_user_form';
 import { getServerSession } from 'next-auth';
 import AccessDenied from '@/app/components/access_denied';
-import authOptions from '@/app/api/auth/authoptions';
+import authOptions from '@/app/api/auth/auth_options';
 
 export default async function EditUser({ params }: { params: { id: string } }) {
   const session = await getServerSession(authOptions);
